Picker code move to the PickerViewClass
