+++ E-Legal/Assets.xcassets/mail_icon.imageset/Contents.json
@@ -2,6 +2,14 @@
   "images" : [
     {
       "idiom" : "universal",
+      "scale" : "1x"
+    },
+    {
+      "idiom" : "universal",
+      "scale" : "2x"
+    },
+    {
+      "idiom" : "universal",
       "filename" : "mail_icon@3x.png",
       "scale" : "3x"
     }
