+++ E-Legal/Assets.xcassets/old-handphone.imageset/Contents.json
@@ -2,6 +2,14 @@
   "images" : [
     {
       "idiom" : "universal",
+      "scale" : "1x"
+    },
+    {
+      "idiom" : "universal",
+      "scale" : "2x"
+    },
+    {
+      "idiom" : "universal",
       "filename" : "old-handphone@3x.png",
       "scale" : "3x"
     }
